services: extract secret link to subservice
