+++ invenio_rdm_records/services/secret_links/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2021 CERN.
+#
+# Invenio-RDM-Records is free software; you can redistribute it and/or modify
+# it under the terms of the MIT License; see LICENSE file for more details.
+
+"""module for Secret Link service."""
+
+from .service import SecretLinkService
+
+__all__ = (
+    "SecretLinkService",
+)
